- Added texture component to the terrain and plane classes.
I'm not happy though, at the moment it works despite the fact that the normal shader only reads in the position and normal. It means that we're sending all of the information to the GPU though in the vertex buffer even if it's not being used. I need to figure out a way to send only the information required, and to be able to change it up and supply each shader with whatever it needs.