- The normal technique should use the full PositionNormalTexture layout. It's only a coincidence that it works with the PositionNormal layout because the Position and Normal are in the same spots. If the PNT layout was instead PTN, then it would fail since the Texture data would be read into the normal spot.
Some stuff I've got together tonight.

Basic questions.

What does the shader want? What does it expect for each vertex
What does the buffer have? Does the geometry support everything that's being asked of it, does it have normals, texcoords, binormals, etc...

If the buffer can't give the shader what it wants, we have a problem. Otherwise for the time being we should just have a single buffer with input and let the shader decide what it does or does not want to read.