Done!
I think it was a camera issue before. The aspect ratio wasn't set when the camera was being built and that was screwing some stuff up and it also pushed the camera forward on the update function.

I've enabled a lot of debug now on the device, and the projects and this gave much more information on what the problem in the shaders and stuff was. Basically the semantics were not named the same.

All looks pretty good now though. Well, I mean... it's a start.